+++ apps/cowswap-frontend/src/modules/tradeFlow/services/swapFlow/index.ts
@@ -15,6 +15,7 @@ import { UiOrderType } from '@cowprotocol/types'
 import { SigningSteps } from 'entities/trade'
 import ms from 'ms.macro'
 import { tradingSdk } from 'tradingSdk/tradingSdk'
+import { sendTransaction } from 'wagmi/actions'
 
 import { PriceImpact } from 'legacy/hooks/usePriceImpact'
 import { partialOrderUpdate } from 'legacy/state/orders/utils'
@@ -29,6 +30,8 @@ import { TradeFlowAnalytics } from 'modules/trade/utils/tradeFlowAnalytics'
 import { getSwapErrorMessage } from 'common/utils/getSwapErrorMessage'
 
 import { TradeFlowContext } from '../../types/TradeFlowContext'
+
+import type { Hex } from 'viem'
 
 const DELAY_BETWEEN_SIGNATURES = ms`500ms`
 
@@ -66,6 +69,7 @@ export async function swapFlow(
   const {
     orderParams,
     context,
+    config,
     permitInfo,
     generatePermitHook,
     permitAmountToSign,
@@ -91,7 +95,7 @@ export async function swapFlow(
     orderParams.appData = await handlePermit({
       appData,
       typedHooks,
-      account,
+      account: account as `0x${string}`,
       inputToken: inputCurrency,
       permitInfo,
       amount: permitAmountToSign,
@@ -177,19 +181,21 @@ export async function swapFlow(
       logTradeFlow('SWAP FLOW', 'STEP 5: presign order (optional)')
       const presignTx = await tradingSdk.getPreSignTransaction({ orderUid: orderId })
 
-      presignTxHash = (
-        await orderParams.signer.sendTransaction(presignTx).catch((error) => {
-          /**
-           * When using Rabby and Safe, the presign transaction is not a real transaction
-           * It's a safe signature
-           */
-          if (error.transactionHash) {
-            return { hash: error.transactionHash }
-          } else {
-            throw error
-          }
-        })
-      ).hash
+      presignTxHash = await sendTransaction(config, {
+        to: presignTx.to as `0x${string}`,
+        value: BigInt(presignTx.value),
+        data: presignTx.data as Hex,
+      }).catch((error) => {
+        /**
+         * When using Rabby and Safe, the presign transaction is not a real transaction
+         * It's a safe signature
+         */
+        if (error.transactionHash) {
+          return error.transactionHash
+        } else {
+          throw error
+        }
+      })
     }
 
     const order = mapUnsignedOrderToOrder({